+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/message/DefaultUIMessageResolver.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.main
+package com.woocommerce.android.ui.message
 
 import android.app.Activity
 import android.view.ViewGroup
@@ -6,10 +6,11 @@ import com.woocommerce.android.R
 import com.woocommerce.android.ui.base.UIMessageResolver
 import javax.inject.Inject
 
-class MainUIMessageResolver @Inject constructor(activity: Activity) : UIMessageResolver {
+class DefaultUIMessageResolver @Inject constructor(activity: Activity) : UIMessageResolver {
     override val snackbarRoot: ViewGroup by lazy {
         requireNotNull(activity.findViewById(R.id.snack_root)) {
             "To be able to use UIMessageResolver, the activity has to contain a Layout with id snack_root"
         }
     }
+    override var anchorViewId: Int? = null
 }