+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/orders/creation/OrderCreateEditRepositoryTest.kt
@@ -3,21 +3,30 @@ package com.woocommerce.android.ui.orders.creation
 import com.woocommerce.android.analytics.AnalyticsEvent
 import com.woocommerce.android.analytics.AnalyticsTracker
 import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
+import com.woocommerce.android.model.Order
 import com.woocommerce.android.tools.SelectedSite
+import com.woocommerce.android.ui.orders.OrderTestUtils
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.junit.Before
 import org.junit.Test
+import org.mockito.kotlin.any
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.eq
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
+import org.mockito.kotlin.whenever
 import org.wordpress.android.fluxc.model.SiteModel
+import org.wordpress.android.fluxc.model.WCOrderStatusModel
+import org.wordpress.android.fluxc.model.order.UpdateOrderRequest
+import org.wordpress.android.fluxc.model.plugin.SitePluginModel
 import org.wordpress.android.fluxc.network.BaseRequest
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooError
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooErrorType
 import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooResult
+import org.wordpress.android.fluxc.network.rest.wpcom.wc.order.CoreOrderStatus
 import org.wordpress.android.fluxc.store.OrderUpdateStore
+import org.wordpress.android.fluxc.store.WooCommerceStore
 import java.math.BigDecimal
 
 @ExperimentalCoroutinesApi
@@ -30,23 +39,27 @@ class OrderCreateEditRepositoryTest : BaseUnitTest() {
     private lateinit var trackerWrapper: AnalyticsTrackerWrapper
     private lateinit var orderUpdateStore: OrderUpdateStore
     private lateinit var selectedSite: SelectedSite
+    private lateinit var wooCommerceStore: WooCommerceStore
+
+    private val defaultSiteModel = SiteModel()
 
     @Before
     fun setUp() {
         trackerWrapper = mock()
 
-        val siteModel = SiteModel()
         selectedSite = mock {
-            on { get() } doReturn siteModel
+            on { get() } doReturn defaultSiteModel
         }
 
         orderUpdateStore = mock {
             onBlocking {
-                createSimplePayment(eq(siteModel), eq("1"), eq(true), eq(null))
+                createSimplePayment(eq(defaultSiteModel), eq("1"), eq(true), eq(null))
             } doReturn WooResult(
                 WooError(WooErrorType.API_ERROR, BaseRequest.GenericErrorType.NETWORK_ERROR, DEFAULT_ERROR_MESSAGE)
             )
         }
+
+        wooCommerceStore = mock()
 
         sut = OrderCreateEditRepository(
             selectedSite = selectedSite,
@@ -54,7 +67,7 @@ class OrderCreateEditRepositoryTest : BaseUnitTest() {
             orderUpdateStore = orderUpdateStore,
             orderMapper = mock(),
             dispatchers = coroutinesTestRule.testDispatchers,
-            wooCommerceStore = mock(),
+            wooCommerceStore = wooCommerceStore,
             analyticsTrackerWrapper = trackerWrapper
         )
     }
@@ -70,5 +83,65 @@ class OrderCreateEditRepositoryTest : BaseUnitTest() {
                 AnalyticsTracker.KEY_FLOW to AnalyticsTracker.VALUE_SIMPLE_PAYMENTS_FLOW
             )
         )
+    }
+
+    @Test
+    fun `when AUTO_DRAFT is not supported then status is changed to PENDING`() = testBlocking {
+        // Given a site using a version that doesn't support AUTO_DRAFT
+        whenever(wooCommerceStore.getSitePlugin(selectedSite.get(), WooCommerceStore.WooPlugin.WOO_CORE))
+            .thenReturn(SitePluginModel().apply { version = "6.2.0" })
+        whenever(orderUpdateStore.createOrder(any(), any()))
+            .thenReturn(WooResult(OrderTestUtils.generateOrder()))
+
+        val order = Order.EMPTY.copy(
+            id = 0L,
+            status = Order.Status.Custom(Order.Status.AUTO_DRAFT)
+        )
+
+        // When the createOrUpdateDraft method is call
+        sut.createOrUpdateDraft(order)
+
+        // Then the order status is changed to PENDING
+        val request = UpdateOrderRequest(
+            status = WCOrderStatusModel(CoreOrderStatus.PENDING.value),
+            lineItems = emptyList(),
+            shippingAddress = null,
+            billingAddress = null,
+            customerNote = order.customerNote,
+            shippingLines = emptyList(),
+            feeLines = emptyList()
+        )
+
+        verify(orderUpdateStore).createOrder(defaultSiteModel, request)
+    }
+
+    @Test
+    fun `when AUTO_DRAFT is supported then status is keep as AUTO_DRAFT`() = testBlocking {
+        // Given a site using a version that support AUTO_DRAFT
+        whenever(wooCommerceStore.getSitePlugin(selectedSite.get(), WooCommerceStore.WooPlugin.WOO_CORE))
+            .thenReturn(SitePluginModel().apply { version = OrderCreateEditRepository.AUTO_DRAFT_SUPPORTED_VERSION })
+        whenever(orderUpdateStore.createOrder(any(), any()))
+            .thenReturn(WooResult(OrderTestUtils.generateOrder()))
+
+        val order = Order.EMPTY.copy(
+            id = 0L,
+            status = Order.Status.Custom(Order.Status.AUTO_DRAFT)
+        )
+
+        // When the createOrUpdateDraft method is call
+        sut.createOrUpdateDraft(order)
+
+        // Then the order status is not changed
+        val request = UpdateOrderRequest(
+            status = WCOrderStatusModel(Order.Status.AUTO_DRAFT),
+            lineItems = emptyList(),
+            shippingAddress = null,
+            billingAddress = null,
+            customerNote = order.customerNote,
+            shippingLines = emptyList(),
+            feeLines = emptyList()
+        )
+
+        verify(orderUpdateStore).createOrder(defaultSiteModel, request)
     }
 }